+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ============
 
 * (feature) Add generic `RichTextTransformer`.
+* (bug) Add `LinkMarksRichTextTransformer` and replace custom implementation, to cover more cases.
 
 
 3.0.0-beta.2